// apps/plano-aee/src/types/interactiveResources.types.ts
export type ActivityType = 
  | 'game' 
  | 'quiz' 
  | 'simulation' 
  | 'story' 
  | 'exercise'
  | 'other';

export type ActivityUsageContext = 
  | 'individual_aee' 
  | 'group_aee' 
  | 'co_teaching' 
  | 'homework'
  | 'assessment';